backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
pandas==1.3.3
numpy==1.21.2
yfinance==0.2.18
oandapyV20==0.7.2
python-jose==3.3.0
python-multipart==0.0.5

